Video 7: SEO con React y Deploy integrando con GitHub; useSEO para añadir también la description.
