+++ src/pages/SearchResults/index.js
@@ -4,7 +4,7 @@ import ListOfGifs from 'componentes/ListOfGifs';
 import { useGifs } from 'hooks/useGifs';
 import useNearScreen from 'hooks/useNearScreen';
 import debounce from 'just-debounce-it'
-import useTitle from 'hooks/useTitle';
+import useTitle from 'hooks/useSEO';
 
 
 export default function SearchResults({params}){
